+++ CHANGES.md
@@ -1,6 +1,29 @@
 # Algebird #
 
+### Version 0.11.0 ###
+* Move CMSHasherByteArray from scalding: https://github.com/twitter/algebird/pull/467
+* Upgrade sbt launcher script (sbt-extras): https://github.com/twitter/algebird/pull/469
+* Create case class macros for algebraic structures: https://github.com/twitter/algebird/pull/466
+* Refactor MapAggregator: https://github.com/twitter/algebird/pull/462
+* Algebird support for spark: https://github.com/twitter/algebird/pull/397
+* Add MapAggregator from 1 (key, aggregator) pair: https://github.com/twitter/algebird/pull/452
+* Remove unnecessary use of scala.math: https://github.com/twitter/algebird/pull/455
+* Don&#39;t call deprecated HyperLogLog methods in tests: https://github.com/twitter/algebird/pull/456
+* Update product_generators.rb: https://github.com/twitter/algebird/pull/457
+* Pzheng/gaussian euclidean: https://github.com/twitter/algebird/pull/448
+
+### Version 0.10.2 ###
+* QTree quantileBounds assert percentile <= 1.0 #447
+
+### Version 0.10.1 ###
+* Make HLL easier to use, add Hash128 typeclass #440
+* add ! to ApproximateBoolean #442
+* add QTreeAggregator and add approximatePercentileBounds to Aggregator #443
+* Make level configurable in QTreeAggregators #444
+
 ### Version 0.10.0 ###
+* HyperLogLogSeries #295
+* CMS: add contramap to convert CMS[K] to CMS[L], add support for String and Bytes, remove Ordering context bound for K #399
 * EventuallyAggregator and variants #407
 * Add MultiAggregator.apply #408
 * Return a MonoidAggregator from MultiAggregator when possible #409